// typechain-types/ClearFundFactory.ts
export interface ClearFundFactoryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "createCampaign"
      | "deployedCampaigns"
      | "getDeployedCampaigns"
      | "platformAdmin"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "CampaignCreated"): EventFragment;

  encodeFunctionData(
    functionFragment: "createCampaign",
    values: [string, string, string, BigNumberish, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "deployedCampaigns",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getDeployedCampaigns",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "platformAdmin",
    values?: undefined
  ): string;

  decodeFunctionResult(
    functionFragment: "createCampaign",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "deployedCampaigns",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getDeployedCampaigns",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "platformAdmin",
    data: BytesLike
  ): Result;
}

export namespace CampaignCreatedEvent {
  export type InputTuple = [
    campaignAddress: AddressLike,
    creator: AddressLike,
    title: string,
    category: string,
    imageIPFS: string,
    goal: BigNumberish,
    deadline: BigNumberish
  ];
  export type OutputTuple = [
    campaignAddress: string,
    creator: string,
    title: string,
    category: string,
    imageIPFS: string,
    goal: bigint,
    deadline: bigint
  ];
  export interface OutputObject {
    campaignAddress: string;
    creator: string;
    title: string;
    category: string;
    imageIPFS: string;
    goal: bigint;
    deadline: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface ClearFundFactory extends BaseContract {
  connect(runner?: ContractRunner | null): ClearFundFactory;
  waitForDeployment(): Promise<this>;

  interface: ClearFundFactoryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  createCampaign: TypedContractMethod<
    [
      _title: string,
      _category: string,
      _imageIPFS: string,
      _goal: BigNumberish,
      _durationInDays: BigNumberish
    ],
    [void],
    "nonpayable"
  >;

  deployedCampaigns: TypedContractMethod<
    [arg0: BigNumberish],
    [string],
    "view"
  >;

  getDeployedCampaigns: TypedContractMethod<[], [string[]], "view">;

  platformAdmin: TypedContractMethod<[], [string], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "createCampaign"
  ): TypedContractMethod<
    [
      _title: string,
      _category: string,
      _imageIPFS: string,
      _goal: BigNumberish,
      _durationInDays: BigNumberish
    ],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "deployedCampaigns"
  ): TypedContractMethod<[arg0: BigNumberish], [string], "view">;
  getFunction(
    nameOrSignature: "getDeployedCampaigns"
  ): TypedContractMethod<[], [string[]], "view">;
  getFunction(
    nameOrSignature: "platformAdmin"
  ): TypedContractMethod<[], [string], "view">;

  getEvent(
    key: "CampaignCreated"
  ): TypedContractEvent<
    CampaignCreatedEvent.InputTuple,
    CampaignCreatedEvent.OutputTuple,
    CampaignCreatedEvent.OutputObject
  >;

  filters: {
    "CampaignCreated(address,address,string,string,string,uint256,uint256)": TypedContractEvent<
      CampaignCreatedEvent.InputTuple,
      CampaignCreatedEvent.OutputTuple,
      CampaignCreatedEvent.OutputObject
    >;
    CampaignCreated: TypedContractEvent<
      CampaignCreatedEvent.InputTuple,
      CampaignCreatedEvent.OutputTuple,
      CampaignCreatedEvent.OutputObject
    >;
  };
}